app/requirements.txt
Django
django-cors-headers
djangorestframework
eyed3
psycopg2-binary
python-dotenv
urllib3
psycopg2-binary
